Add profile template
With standardised icons and order of links & a brief instructions in the template

The order is roughly based on the spectrum of "about me (i.e a profile on personal
or institutional website) to "get in contact" (i.e endig with e-mail preceeded
by twitter/mastodon).

In between more "scholarly" or established sites have been given priority,
viz. they are placed higher in the list -> will appear first/leftmost on
rendered profiles.

Still, the assessment of what is most "scholarly" or "established" is based
on a largely subjective assessment, so the order is fairly arbitrary --
nevertheless, the order itself is not the point: it is the standardization.

Added "started" and "ended" metadata to ease the transisition of members.